requirements.txt
asgiref==3.6.0
Django==4.1.7
django-choices==1.7.2
djangorestframework==3.14.0
gevent==22.10.2
greenlet==2.0.2
gunicorn==20.1.0
PyJWT==2.6.0
pytz==2022.7.1
six==1.16.0
sqlparse==0.4.3
zope.event==4.6
zope.interface==5.5.2
